Conflicts:
	statistik-web/src/main/java/se/inera/statistics/web/service/ChartDataService.java
	statistik-web/src/main/java/se/inera/statistics/web/service/ProtectedChartDataService.java
	statistik-web/src/main/webapp/js/app/services/businessFilter.js

+++ statistik-service/src/test/java/se/inera/statistics/service/queue/ReceiverIntegrationTest.java
@@ -69,12 +69,12 @@ public class ReceiverIntegrationTest {
         assertEquals(12, webData.getRows().size());
 
         for (int i = 0; i < 3; i++) {
-            assertEquals(0, webData.getRows().get(i).getFemale().intValue());
-            assertEquals(2, webData.getRows().get(i).getMale().intValue());
+            assertEquals(0, webData.getRows().get(i).getFemale());
+            assertEquals(2, webData.getRows().get(i).getMale());
         }
         for (int i = 3; i < 12; i++) {
-            assertEquals(0, webData.getRows().get(i).getFemale().intValue());
-            assertEquals(0, webData.getRows().get(i).getMale().intValue());
+            assertEquals(0, webData.getRows().get(i).getFemale());
+            assertEquals(0, webData.getRows().get(i).getMale());
         }
 
         assertEquals(2, wideLine.count());
